chore(swift): Upgrade gophercloud dependencies
This CL upgrades the gophercloud used by the swift backend to the latest
version. Only a single parameter needs to be converted from int to
int64, which is safe.

+++ internal/backend/remote-state/swift/client.go
@@ -288,7 +288,7 @@ func (c *RemoteClient) put(object string, data []byte, deleteAfter int, ifNoneMa
 	}
 
 	if deleteAfter >= 0 {
-		createOpts.DeleteAfter = deleteAfter
+		createOpts.DeleteAfter = int64(deleteAfter)
 	}
 
 	if ifNoneMatch != "" {
